backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose
python-multipart
email-validator

celery
redis

pandas
numpy
matplotlib
yfinance

requests
beautifulsoup4

reportlab
alembic
